+++ tests/PHPStan/Rules/Methods/FinalPrivateMethodRuleTest.php
@@ -2,9 +2,9 @@
 
 namespace PHPStan\Rules\Methods;
 
+use PHPStan\Php\PhpVersion;
 use PHPStan\Rules\Rule;
 use PHPStan\Testing\RuleTestCase;
-use function array_merge;
 use const PHP_VERSION_ID;
 
 /** @extends RuleTestCase<FinalPrivateMethodRule> */
@@ -16,34 +16,60 @@ class FinalPrivateMethodRuleTest extends RuleTestCase
 		return new FinalPrivateMethodRule();
 	}
 
-	public function testRule(): void
+	public function dataRule(): array
 	{
-		$errors = [];
-		if (PHP_VERSION_ID >= 80000) {
-			$errors = [
+		return [
+			[
+				70400,
+				[],
+			],
+			[
+				80000,
 				[
-					'Private method FinalPrivateMethod\Foo::foo() cannot be final as it is never overridden by other classes.',
-					8,
+					[
+						'Private method FinalPrivateMethod\Foo::foo() cannot be final as it is never overridden by other classes.',
+						8,
+					],
 				],
-			];
+			],
+		];
+	}
+
+	/**
+	 * @dataProvider dataRule
+	 * @param list<array{0: string, 1: int, 2?: string}> $errors
+	 */
+	public function testRule(int $phpVersion, array $errors): void
+	{
+		$testVersion = new PhpVersion($phpVersion);
+		$runtimeVersion = new PhpVersion(PHP_VERSION_ID);
+
+		if (
+			$testVersion->getMajorVersionId() !== $runtimeVersion->getMajorVersionId()
+			|| $testVersion->getMinorVersionId() !== $runtimeVersion->getMinorVersionId()
+		) {
+			$this->markTestSkipped('Test requires PHP version ' . $phpVersion);
 		}
 
-		$errors = array_merge($errors, [
+		$this->analyse([__DIR__ . '/data/final-private-method.php'], $errors);
+	}
+
+	public function testRulePhpVersions(): void
+	{
+		$this->analyse([__DIR__ . '/data/final-private-method-phpversions.php'], [
 			[
-				'Private method FinalPrivateMethod\FooBarPhp8orHigher::foo() cannot be final as it is never overridden by other classes.',
+				'Private method FinalPrivateMethodPhpVersions\FooBarPhp8orHigher::foo() cannot be final as it is never overridden by other classes.',
+				9,
+			],
+			[
+				'Private method FinalPrivateMethodPhpVersions\FooBarPhp74OrHigher::foo() cannot be final as it is never overridden by other classes.',
+				29,
+			],
+			[
+				'Private method FinalPrivateMethodPhpVersions\FooBarBaz::foo() cannot be final as it is never overridden by other classes.',
 				39,
 			],
-			[
-				'Private method FinalPrivateMethod\FooBarPhp74OrHigher::foo() cannot be final as it is never overridden by other classes.',
-				59,
-			],
-			[
-				'Private method FinalPrivateMethod\FooBarBaz::foo() cannot be final as it is never overridden by other classes.',
-				69,
-			],
 		]);
-
-		$this->analyse([__DIR__ . '/data/final-private-method.php'], $errors);
 	}
 
 }
